requirements.txt
PyNaCl
requests
PyYAML